updated detil

+++ database/migrations/2022_09_04_145336_create_dupak_table.php
@@ -18,7 +18,7 @@ class CreateDupakTable extends Migration
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('butir_kegiatan_id');
             $table->integer('volume');
-            $table->integer('angka_kredit_usulan');
+            $table->decimal('angka_kredit_usulan',4,3);
             $table->date('tanggal_pelaksanaan');
             $table->string('evidence')->nullable();
             $table->timestamps();
